% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapstat_metadata.R
\name{check_metadata}
\alias{check_metadata}
\title{Check mapstat metadata}
\usage{
check_metadata(meta_df)
}
\arguments{
\item{meta_df}{Metadata data frame that should be checked. It is expected that
you imported this one with \code{\link{get_multiple_metadata}}.}
}
\value{
Nothing is returned. Only prints a summary the console
}
\description{
Check mapstat metadata
}
