% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapstat_metadata.R
\name{get_multiple_metadata}
\alias{get_multiple_metadata}
\title{Import the metadata from many mapstat files}
\usage{
get_multiple_metadata(input, type = "directory", ...)
}
\arguments{
\item{input}{Either the name of a directory containing the files to be
imported, or a character vector containing file paths.}

\item{type}{Specify whether the \code{input} is of type "directory" (default)
or "vector".}

\item{...}{Additional arguments to be passed to \code{\link{get_single_metadata}}.
At the moment, the only such argument is \code{shorten_sample_id}.}
}
\value{
A data frame containing the metadata of the provided mapstat files.
  A quick summary of each column is printed to the console.
}
\description{
Import the metadata from many mapstat files
}
