% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapstat_metadata.R
\name{get_single_metadata}
\alias{get_single_metadata}
\title{Import the metadata of a mapstat file}
\usage{
get_single_metadata(path, shorten_sample_id = TRUE)
}
\arguments{
\item{path}{Path to file}

\item{shorten_sample_id}{Should the \code{sample_id} variable in the output
  be shortened?

  If \code{TRUE}, "samplename" is returned instead of the full path
  "foo/bar/samplename.mapstat". If\code{FALSE}, the full path is returned.}
}
\value{
A list containing the six metadata fields parsed from the 'mapstat' file.
}
\description{
Import the metadata of a mapstat file
}
