% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapstat_import.R
\name{read_multiple_mapstats}
\alias{read_multiple_mapstats}
\title{Import multiple mapstat files}
\usage{
read_multiple_mapstats(input, type = "directory", ...)
}
\arguments{
\item{input}{Either the name of a directory containing the files to be
imported, or a character vector containing file paths.}

\item{type}{Specify whether the \code{input} is of type "directory" (default)
or "vector".}

\item{...}{Additional arguments to be passed to \code{\link{read_single_mapstat}},
auch as \code{shorten_sample_id} and \code{engine}.}
}
\value{
A data frame containing information from all imported mapstat files.
  The (by default shortened) filenames are added in a column called \code{sample_id}.
}
\description{
Import multiple mapstat files
}
