% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refdata_import.R
\name{read_refdata}
\alias{read_refdata}
\title{Import a refdata file}
\usage{
read_refdata(path)
}
\arguments{
\item{path}{Path to refdata file}
}
\value{
A data frame containing the refdata information.
}
\description{
Import a refdata file
}
