% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapstat_import.R
\name{read_single_mapstat}
\alias{read_single_mapstat}
\title{Import a mapstat file}
\usage{
read_single_mapstat(path, shorten_sample_id = TRUE)
}
\arguments{
\item{path}{Path to file}

\item{shorten_sample_id}{Should the \code{sample_id} variable in the output
  be shortened?

  If \code{TRUE}, "samplename" is returned instead of the full path
  "foo/bar/samplename.mapstat". If\code{FALSE}, the full path is returned.}
}
\description{
Import a mapstat file
}
